+++ modules/default/weather/providers/ukmetoffice.js
@@ -1,12 +1,16 @@
 /* global WeatherProvider, WeatherObject, WeatherUtils */
 
-/* This class is a provider for UK Met Office Datapoint,
+/*
+ * This class is a provider for UK Met Office Datapoint,
  * see https://www.metoffice.gov.uk/
  */
 WeatherProvider.register("ukmetoffice", {
-	// Set the name of the provider.
-	// This isn't strictly necessary, since it will fallback to the provider identifier
-	// But for debugging (and future alerts) it would be nice to have the real name.
+
+	/*
+	 * Set the name of the provider.
+	 * This isn't strictly necessary, since it will fallback to the provider identifier
+	 * But for debugging (and future alerts) it would be nice to have the real name.
+	 */
 	providerName: "UK Met Office",
 
 	// Set the default config properties that is specific to this provider
@@ -21,8 +25,11 @@ WeatherProvider.register("ukmetoffice", {
 		this.fetchData(this.getUrl("3hourly"))
 			.then((data) => {
 				if (!data || !data.SiteRep || !data.SiteRep.DV || !data.SiteRep.DV.Location || !data.SiteRep.DV.Location.Period || data.SiteRep.DV.Location.Period.length === 0) {
-					// Did not receive usable new data.
-					// Maybe this needs a better check?
+
+					/*
+					 * Did not receive usable new data.
+					 * Maybe this needs a better check?
+					 */
 					return;
 				}
 
@@ -42,8 +49,11 @@ WeatherProvider.register("ukmetoffice", {
 		this.fetchData(this.getUrl("daily"))
 			.then((data) => {
 				if (!data || !data.SiteRep || !data.SiteRep.DV || !data.SiteRep.DV.Location || !data.SiteRep.DV.Location.Period || data.SiteRep.DV.Location.Period.length === 0) {
-					// Did not receive usable new data.
-					// Maybe this needs a better check?
+
+					/*
+					 * Did not receive usable new data.
+					 * Maybe this needs a better check?
+					 */
 					return;
 				}
 
@@ -86,8 +96,11 @@ WeatherProvider.register("ukmetoffice", {
 			if (periodDate.isSameOrAfter(moment.utc().startOf("day"))) {
 				// check this is the period we want, after today the diff will be -ve
 				if (moment().diff(periodDate, "minutes") > 0) {
-					// loop round the reports looking for the one we are in
-					// $ value specifies the time in minutes-of-the-day: 0, 180, 360,...1260
+
+					/*
+					 * loop round the reports looking for the one we are in
+					 * $ value specifies the time in minutes-of-the-day: 0, 180, 360,...1260
+					 */
 					for (const rep of period.Rep) {
 						const p = rep.$;
 						if (timeInMins >= p && timeInMins - 180 < p) {
@@ -117,8 +130,10 @@ WeatherProvider.register("ukmetoffice", {
 	generateWeatherObjectsFromForecast (forecasts) {
 		const days = [];
 
-		// loop round the (5) periods getting the data
-		// for each period array, Day is [0], Night is [1]
+		/*
+		 * loop round the (5) periods getting the data
+		 * for each period array, Day is [0], Night is [1]
+		 */
 		for (const period of forecasts.SiteRep.DV.Location.Period) {
 			const weather = new WeatherObject();