+++ tests/configs/modules/compliments/compliments_e2e_cron_entry.js
@@ -1,0 +1,18 @@
+let config = {
+	modules: [
+		{
+			module: "compliments",
+			position: "middle_center",
+			config: {
+				specialDayUnique: true,
+				compliments: {
+					anytime: ["just a test"],
+					"* * * * *": ["anytime cron"]
+				}
+			}
+		}
+	]
+};
+
+/*************** DO NOT EDIT THE LINE BELOW ***************/
+if (typeof module !== "undefined") { module.exports = config; }